refactor: utils 컴포넌트 utils 폴더 하위로 이동
